Initial daemon + use bind mount within mzr shells
If two separate mount namespaces are created, and the same overlay
mount command is run in each, the resulting filesystems do not behave
consistently. The solution here is to have a daemon process which

Even if there is another way to work around this problem, there are
quite a few potential future uses for such a daemon.